+++ src/features/flags/flags.ts
@@ -2,7 +2,7 @@ import type { App } from "electron";
 
 import { flags } from "../../constants/flags";
 import { settings } from "../../constants/settings";
-import { settingsStore } from "../../scripts/settings";
+import { settingsStore } from "../../scripts/settingsStore";
 import { Logger } from "../logger";
 
 /**