+++ src/scripts/tray.ts
@@ -4,7 +4,7 @@ import { type BrowserWindow, Tray } from "electron";
 import { settings } from "../constants/settings";
 import { SUPPORTED_TRAY_ICON_EXTENSIONS } from "../constants/trayIcon";
 import { getMenu } from "./menu";
-import { settingsStore } from "./settings";
+import { settingsStore } from "./settingsStore";
 
 let tray: Tray;